fix: changes convites to user email from username
